requirements.txt
customtkinter
pillow
pywin32
icoextract